--- crates/tools/bindings/src/core_com.txt
@@ -1,18 +1,0 @@
---out crates/libs/core/src/imp/com_bindings.rs
---flat
-
---filter
-    AGILEREFERENCE_DEFAULT
-    CO_E_NOTINITIALIZED
-    E_INVALIDARG
-    E_NOINTERFACE
-    E_POINTER
-    IAgileObject
-    IAgileReference
-    IWeakReference
-    IWeakReferenceSource
-    JSCRIPT_E_CANTEXECUTE
-    REGDB_E_CLASSNOTREG
-    RoGetAgileReference
-    RPC_E_DISCONNECTED
-    S_OK
